pyproject.toml
[project]
name = "meeting-assistant"
version = "0.1.0"
description = "Realtime speech recognition, translation, and summarization using Gemini and Agno"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.5.11",
    "fastmcp>=3.1.1",
    "flask-socketio>=5.6.1",
    "flask>=3.1.3",
    "google-genai>=1.69.0",
    "numpy>=2.4.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sounddevice>=0.5.5",
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
]
